cluster_rl_next/requirements_cluster.txt
absl-py==2.1.0
aiosignal==1.4.0
asttokens==2.4.1
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.0
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.9
einops==0.8.0
executing==2.1.0
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.8.0
fsspec==2024.9.0
grpcio==1.67.0
gymnasium==0.28.1
hnswlib==0.8.0
idna==3.11
imageio==2.36.0
ipython==8.28.0
jax-jumpy==1.0.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
lazy_loader==0.4
lz4==4.4.4
Markdown==3.7
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.2
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.1
numpy==1.26.4
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
protobuf==5.28.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==12.0.1
pyboy==2.4.0
Pygments==2.18.0
pyparsing==3.2.0
PySDL2==0.9.16
pysdl2-dll==2.30.2
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.3
ray==2.9.0
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
scikit-image==0.24.0
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
stable_baselines3==2.3.2
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tifffile==2024.9.20
torch==2.5.0
traitlets==5.14.3
typer==0.20.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.5.0
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.4
wrapt==2.0.0
